+++ packages/sdk/src/resolve/resolve-accounts.js
@@ -1,8 +1,6 @@
 import {invariant} from "@onflow/util-invariant"
-import {
-  isTransaction,
-  createSignableVoucher,
-} from "../interaction/interaction.js"
+import {isTransaction} from "../interaction/interaction.js"
+import {createVoucher} from "../utils"
 
 const isFn = v => typeof v === "function"
 export function buildPreSignable(acct, ix) {
@@ -15,7 +13,7 @@ export function buildPreSignable(acct, ix) {
       args: ix.message.arguments.map(d => ix.arguments[d].asArgument),
       data: {},
       interaction: ix,
-      voucher: createSignableVoucher(ix),
+      voucher: createVoucher(ix),
     }
   } catch (error) {
     console.error("buildPreSignable", error)